requirements.txt
customtkinter
pillow
bcrypt
ctktooltip
pyinstaller